net 1

+++ atlas-core/src/net/rpc/packet_request.rs
@@ -1,14 +1,13 @@
 use bytes::Bytes;
-use serde::{Deserialize, Serialize};
 use serde::de::DeserializeOwned;
-
+use serde::{Deserialize, Serialize};
 
 pub type AtlasRawRequest = AtlasWireRequest<Bytes>;
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct AtlasWireRequest<T> {
     pub id: u64,
-    pub slot_index: usize,
+    pub slot_index: u64,
     pub method: u32,
     pub payload: T,
 }
@@ -18,8 +17,7 @@ where
     T: Serialize + DeserializeOwned,
 {
     pub fn from_raw(raw: AtlasRawRequest) -> Result<Self, String> {
-        let payload = rmp_serde::from_slice(&raw.payload.as_ref())
-            .map_err(|e| e.to_string())?;
+        let payload = rmp_serde::from_slice(&raw.payload.as_ref()).map_err(|e| e.to_string())?;
 
         Ok(Self {
             id: raw.id,
@@ -30,8 +28,7 @@ where
     }
 
     pub fn into_raw(self) -> Result<AtlasRawRequest, String> {
-        let payload = rmp_serde::to_vec(&self.payload)
-            .map_err(|e| e.to_string())?;
+        let payload = rmp_serde::to_vec(&self.payload).map_err(|e| e.to_string())?;
 
         Ok(AtlasRawRequest {
             id: self.id,
